# Problem: Finding replay_lsn with lsn tracking is 7.5 times slower