Currently, we maintain a LSN mapping which tracks the target insert LSN against current DML operation's LSN. This mapping will be used to determine the replay_lsn. The mapping is implemented using LIST and every insert executes `pg_current_wal_insert_lsn` procedure on target server. When there is a need find current replay_lsn, another procedure `pg_current_wal_flush_lsn` executed on target and respective source LSN is determinted by iterating the mapping LIST. Additionally, we persist the list into a sqlite table on each sentinel sync leading more delay due to local disk write.

Solution: Use replication origin progress lsn as replay_lsn

We create postgres replication origin on target and use it to track the apply progress which includes source operation LSN. We can simply reuse the origin progress tracking to find the replay_lsn. By doing this, we eliminate a Postgres procedure call on each transaction and bunch of local disk write by sqlite. Removing the above mentioned functionality improves the overall write throughput by ~7.5 times.

### Without fix(main):
```
time pgcopydb stream apply --resume --not-consistent ~/dimitri/bench.sql
09:48:19.678 112637 INFO   Running pgcopydb version 0.15.38.gc1bc3d9 from "/home/ubuntu/dimitri/pgcopydb/src/bin/pgcopydb/pgcopydb"
09:48:19.726 112637 INFO   Using work dir "/tmp/pgcopydb"
09:48:19.793 112637 INFO   Setting up previous LSN from replication origin "pgcopydb" progress at 0/0
09:48:19.868 112637 INFO   Replaying changes from file "/home/ubuntu/dimitri/bench.sql"
pgcopydb stream apply --resume --not-consistent ~/dimitri/bench.sql  9.25s user 11.48s system 7% cpu 4:54.31 total
```
122987 / 4m54s = 419 records/sec

### With fix:
```
time pgcopydb stream apply --resume --not-consistent ~/dimitri/bench.sql
10:01:42.701 113419 INFO   Running pgcopydb version 0.15.38.gc1bc3d9 from "/home/ubuntu/dimitri/pgcopydb/src/bin/pgcopydb/pgcopydb"
10:01:42.750 113419 INFO   Using work dir "/tmp/pgcopydb"
10:01:42.814 113419 INFO   Setting up previous LSN from replication origin "pgcopydb" progress at 0/0
10:01:42.889 113419 INFO   Replaying changes from file "/home/ubuntu/dimitri/bench.sql"
pgcopydb stream apply --resume --not-consistent ~/dimitri/bench.sql  13.22s user 5.46s system 47% cpu 39.205 total
```
122987 / 39.2s = 3,137 records/sec

### TODO

- [x] Remove lsn_tracking functionality