@dimitri 2024-04-02
I'm not sure about your work here. The idea is that we are using `{ "synchronous_commit", "off" }` when applying changes on the target, in order to avoid waiting for the disk flush at each transaction commit. Then we can't take the replay_lsn (nor the origin lsn) at face value, because at COMMIT time we did not wait to make sure the replayed transaction was indeed committed to disk yet.

Our code is a little complex around this, the goal is to report a trustworthy replay_lsn to the source system and replication slot, in order to ensure we can still replay any failed transaction later. 

Maybe you could try removing all the code around the LSN mapping and using `{ "synchronous_commit", "on" }` in the replay connection and see what are the performance implications of that?