@arajkumar 2024-04-02
> I'm not sure about your work here. The idea is that we are using `{ "synchronous_commit", "off" }` when applying changes on the target, in order to avoid waiting for the disk flush at each transaction commit. Then we can't take the replay_lsn (nor the origin lsn) at face value, because at COMMIT time we did not wait to make sure the replayed transaction was indeed committed to disk yet.

@dimitri Since we read the origin progress using a separate postgres connection, isn't the origin progress LSN representing a txn which is committed to disk?