@dimitri 2024-04-02
> @dimitri Since we read the origin progress using a separate postgres connection, isn't the origin progress LSN representing a txn which is committed to disk?

So... the origin LSN is valid in the context of the source database. The commit-to-disk aspect is asked in the context of the target database, where the transaction has its own (different) LSN. The replay transaction is associated with a source LSN and happens on the target database with `synchronous_commit` set to `off`, which means that when the COMMIT query is sent and the code receives an answer, we don't know yet if the COMMIT made it safely to disk on the target.

That's why we have this complexity here. We want to make sure to report a replay_lsn to the source database that has already made it safely to disk on the target database, and that needs to be a _source lsn_, not a _target lsn_. The only way to bypass this complexity that I can see is to use `synchronous_commit` set to `on`, which means that the source LSN being replayed is safe to publish as the replay_lsn to the source system right at COMMIT response time (on the target).