@arajkumar 2024-04-03
@dimitri, let's consider a scenario where we're replicating 100 transactions from source to target with `SET synchronous_commit=off`. While sending those 100 transactions, we also update the origin progress with each transaction's commit LSN (from the source). Suppose that, so far, the target has committed up to the 50th transaction to disk. If we run `SELECT pg_replication_origin_progress('pgcopydb', true)` on the target, wouldn't it return the LSN of the 50th transaction, which could then be considered as the replay_lsn?