@dimitri 2024-04-03
> @dimitri, let's consider a scenario where we're replicating 100 transactions from source to target with `SET synchronous_commit=off`. While sending those 100 transactions, we also update the origin progress with each transaction's commit LSN (from the source). Suppose that, so far, the target has committed up to the 50th transaction to disk. If we run `SELECT pg_replication_origin_progress('pgcopydb', true)` on the target, wouldn't it return the LSN of the 50th transaction, which could then be considered as the replay_lsn?

Given Postgres docs for the function, it might work, yes. I am not sure it does though. I do not know how to compare the wording “flushed to disk” with the asynchronous_commit aspects of what we do. What kind of testing did you do around it? Did you study Postgres source code already to better understand the guarantees here?

> Returns the replay location for the given replication origin. The parameter flush determines whether the corresponding local transaction will be guaranteed to have been flushed to disk or not.