@arajkumar 2024-04-05
@dimitri, I spent sometime with the postgres source, as per my understanding, `SELECT pg_replication_origin_progress('pgcopydb', true)` would ensure the local txn is flushed before returning the progress LSN. 

Please refer this code snippet => https://github.com/postgres/postgres/blob/6f132ed693b6a0be0997f092c42abf14878362ac/src/backend/replication/logical/origin.c#L1044-L1045

Additionally, I wrote a simple SQL test to confirm,

## Create replication origin
```
psql "$PG_URI" -f << EOF
SELECT pg_replication_origin_create('pgcopydb_test');
CREATE TABLE IF NOT EXISTS test(id int8);
EOF
```

## Advance replication origin and verify it is always less than flush lsn

```
psql "$PG_URI" -f - << EOF
SELECT pg_replication_origin_session_setup('pgcopydb_test');

BEGIN;
INSERT INTO test VALUES(1);
SELECT pg_current_wal_insert_lsn() AS lsn \gset
SELECT pg_replication_origin_xact_setup(:'lsn'::pg_lsn, now());
INSERT INTO test VALUES(2);

SELECT  pg_replication_origin_progress('pgcopydb_test', true) < pg_current_wal_flush_lsn()  AS flush;
COMMIT;
EOF
```

I had used `pg_current_wal_insert_lsn` as a origin LSN  for the ease of demonstration.