@dimitri 2024-04-05
Thanks for checking on that @arajkumar ! Next step would be to get rid of the whole lsn tracking infrastructure in the code then, including the SQLite catalogs entirely. Can you extend your PR to get rid of all the extra code that's not needed anymore?